requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-groq
langchain-openai
pandas
numpy 
scikit-learn
ipykernel
